Fixing field

+++ Shared/Vicente/script.sql
@@ -5,7 +5,7 @@ CREATE DATABASE measures;
 CREATE TABLE measures(
 id serial PRIMARY KEY,
 transactions numeric,
-accumulate double precision,
+accumulate numeric,
 memory numeric,
 time_elapsed numeric(12,10),
 exec_date timestamp default now());
